[refactor/#24] - 출석 체크 관련 바인딩 어댑터 통합 및 패키지 이동
